statistics: disable parallel to avoid data race when test
