+++ statistics/handle/gc_serial_test.go
@@ -37,7 +37,6 @@ func createTestKitAndDom(t *testing.T) (*testkit.TestKit, *domain.Domain, func()
 }
 
 func TestGCStats(t *testing.T) {
-	t.Parallel()
 	testKit, dom, clean := createTestKitAndDom(t)
 	defer clean()
 	testKit.MustExec("set @@tidb_analyze_version = 1")
@@ -71,7 +70,6 @@ func TestGCStats(t *testing.T) {
 }
 
 func TestGCPartition(t *testing.T) {
-	t.Parallel()
 	testKit, dom, clean := createTestKitAndDom(t)
 	defer clean()
 	testKit.MustExec("set @@tidb_analyze_version = 1")
@@ -110,7 +108,6 @@ func TestGCPartition(t *testing.T) {
 }
 
 func TestGCExtendedStats(t *testing.T) {
-	t.Parallel()
 	testKit, dom, clean := createTestKitAndDom(t)
 	defer clean()
 	testKit.MustExec("set session tidb_enable_extended_stats = on")
@@ -155,7 +152,6 @@ func TestGCExtendedStats(t *testing.T) {
 }
 
 func TestGCColumnStatsUsage(t *testing.T) {
-	t.Parallel()
 	testKit, dom, clean := createTestKitAndDom(t)
 	defer clean()
 	testKit.MustExec("use test")
